+++ src/OpenTelemetry.AutoInstrumentation/Constants.cs
@@ -24,6 +24,11 @@ internal static class Constants
         public const string AttributeHttpResponseHeaderPrefix = "http.response.header";
     }
 
+    public static class ResourceAttributes
+    {
+        public const string AttributeServiceName = "service.name";
+    }
+
     public static class ConfigurationValues
     {
         public const string None = "none";